requirements.txt
Django==2.0.2
django-adminlte2==0.4.1
django-braces==1.13.0
django-debug-toolbar==1.11
django-developer-panel==0.1.2
djangorestframework==3.9.1
Pillow==5.4.1
pytz==2018.9
sqlparse==0.2.4
